requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.5
psycopg2==2.9.10
python-dotenv==1.0.1
Flask-CORS==4.0.0
pytz==2024.1